pyproject.toml
[project]
name = "deep-agent-project"
version = "0.1.0"
description = "Writer-Reviewer multi-agent system based on deepagents SDK"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepagents>=0.4.11",
    "langchain>=1.2.0",
    "langgraph>=1.1.3",
    "langchain-qwq>=0.3.4",
    "langchain-openai>=1.0.0",
    "langchain-anthropic>=0.3.0",
    "langchain-deepseek>=1.0.1",
    "pyyaml>=6.0",
    "tavily-python>=0.5.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0",
    "langchain-mcp-adapters>=0.2.2",
    "mcp>=1.26.0",
    "httpx[socks]>=0.28.1",
    "langchain-openrouter>=0.2.0",
]

[dependency-groups]
dev = ["pytest>=8.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
# 确保 import main 和 from src.xxx import ... 都能工作
pythonpath = ["."]

# Virtual environment naming convention:
#   Windows: UV_PROJECT_ENVIRONMENT=.venv-win
#   WSL:     UV_PROJECT_ENVIRONMENT=.venv-wsl
# Set this env var before running uv commands (sync, run, etc.)
